refactor: edit gatsby-browser
