+++ gatsby/browser/on-route-update.js
@@ -1,0 +1,5 @@
+import setThemeFavicon from "../../src/utils/set-theme-favicon"
+
+const onRouteUpdate = () => setThemeFavicon()
+
+export default onRouteUpdate
